update methods location
